refactor(checker): extract contextual_refresh_args helper to hold resolve_signatures.rs under 2000 LOC
Move the per-overload contextual-refresh argument collector out of the
resolve_overloaded_call_with_signatures monolith into the sibling
overload_resolution/helpers.rs (pure code motion, pub(super)) so
resolve_signatures.rs stays under the 2000-line arch cap (was 2008 after this
PR's TS2772 elaboration threading). syntax_kind_ext import moves with it; no
behavior change.
